Alterações para a versão 2.1.3.3 -------------------------------- - Alterando os lugares onde Cliente e Fornecedor recebem new Integer() para receberem new Long(). - Alterando a maneira de passar a data de cadastro e nascimento para os métodos setDataNascimento e setDataCadastro de Cliente. - Mudanças na library gzutil. - Novo manual de utilização.

+++ src/br/com/gz/migration/datafile/FornecedorDataFile.java
@@ -223,7 +223,7 @@ public class FornecedorDataFile extends DataFile {
 			format = new MercoFlexFormat();
 		}
 
-		f.setCodigo(new Integer(format.toNumeric(o[i++].toString(), false)));
+		f.setCodigo(new Long(format.toNumeric(o[i++].toString(), false)));
 		f.setNomeFantasia(format.toNomeFantasia(o[i++].toString()));
 		f.setRazaoSocial(format.toRazaoSocial(o[i++].toString()));
 		f.setCgc(format.toNumeric(o[i++].toString(), false));
@@ -287,7 +287,7 @@ public class FornecedorDataFile extends DataFile {
 			format = new MercoFlexFormat();
 		}
 
-		f.setCodigo(new Integer(format.toNumeric(o[i++].toString(), false)));
+		f.setCodigo(new Long(format.toNumeric(o[i++].toString(), false)));
 		f.setNomeFantasia(format.toNomeFantasia(o[i++].toString()));
 		f.setRazaoSocial(format.toRazaoSocial(o[i++].toString()));
 		f.setCgc(format.toNumeric(o[i++].toString(), false));
